Fix for case when surfaceDestroyed is never called
On my Nexus S, surfaceDestroyed is never called if the window loses
focus very quickly.  This means that when surfaceCreated is called upon
regaining focus, the GLThread instance still thinks the surface is bad
(mSurfaceIsBad == true).  This fix simply sets mSurfaceIsBad to true in
surfaceCreated since we know the surface should be valid at this point.